project1/requirements.txt
torch
matplotlib
pillow
imageio
